+++ ysqlite3/vfs/page_transforming_file.hpp
@@ -4,6 +4,7 @@
 #include "../error.hpp"
 #include "file.hpp"
 
+#include <array>
 #include <cstdint>
 #include <cstring>
 #include <type_traits>
@@ -37,18 +38,13 @@ public:
 			} else {
 				decode_page(buffer);
 			}
+
+			_parse_parameters(buffer.cast<const std::uint8_t*>(), offset);
 		} else {
 #if PRINT_DEBUG
 			puts("forwarding read");
 #endif
 			Parent::read(buffer, offset);
-		}
-
-		// check reserve size
-		if (this->format == file_format::main_db && offset <= 20 && offset + buffer.size() >= 21) {
-			if (!check_reserve_size(buffer.begin()[20 - offset])) {
-				throw std::system_error{ sqlite3_errc::generic, "bad reserve size" };
-			}
 		}
 	}
 	void write(span<const std::uint8_t*> buffer, sqlite3_int64 offset) override
@@ -57,14 +53,8 @@ public:
 		printf("write to %s: %zi from %lli\n", name_of(this->format), buffer.size(), offset);
 #endif
 
-		// check reserve size
-		if (this->format == file_format::main_db && offset <= 20 && offset + buffer.size() >= 21) {
-			if (!check_reserve_size(buffer.begin()[20 - offset])) {
-				throw std::system_error{ sqlite3_errc::generic, "bad reserve size" };
-			}
-		}
-
 		if (_is_page(buffer.size(), offset)) {
+			_parse_parameters(buffer, offset);
 			_tmp_buffer.resize(buffer.size());
 			std::memcpy(_tmp_buffer.data(), buffer.begin(), _tmp_buffer.size());
 
@@ -94,12 +84,31 @@ protected:
 
 private:
 	std::vector<std::uint8_t> _tmp_buffer;
+	std::uint32_t _page_size = 0;
 
 	bool _is_page(std::size_t size, sqlite3_int64 offset) const noexcept
 	{
 		return (this->format == file_format::main_db /* ||
 		        (this->format == file_format::main_journal && offset) */) &&
 		       size >= 512;
+	}
+	void _parse_parameters(span<const std::uint8_t*> buffer, sqlite3_int64 offset)
+	{
+		if (this->format == file_format::main_db) {
+			// read page size
+			if (offset <= 16 && offset + buffer.size() >= 18) {
+				_page_size = buffer.begin()[16 - offset] << 8 | buffer.begin()[17 - offset];
+				if (_page_size == 1) {
+					_page_size = 65536;
+				}
+			}
+
+			// check reserve size
+			if (offset <= 20 && offset + buffer.size() >= 21 &&
+			    !check_reserve_size(buffer.begin()[20 - offset])) {
+				throw std::system_error{ sqlite3_errc::generic, "bad reserve size" };
+			}
+		}
 	}
 };